Move ReservationCalendarControls to screens
Refs #435.
